Updated test for creating an Owner. - Updated test to follow coding standards, by adding a header, a setUp() method and easier test methods.
