+++ tests/selenium_tests/phpunit.xml
@@ -9,6 +9,7 @@
             <file>shared_tests/create_user_curator.php</file>
             <file>shared_tests/assign_curator_to_IVD.php</file>
             <file>shared_tests/create_user_collaborator.php</file>
+            <file>shared_tests/create_user_owner.php</file>
             <file>shared_tests/create_user_submitter.php</file>
             <file>shared_tests/create_disease_IVA.php</file>
             <file>shared_tests/convert_LOVD2_export_IVD.php</file>
@@ -48,7 +49,7 @@
             <file>shared_tests/create_user_curator.php</file>
             <file>shared_tests/create_user_collaborator.php</file>
             <file>shared_tests/create_user_submitter.php</file>
-            <file>authorization_tests/create_user_owner.php</file>
+            <file>shared_tests/create_user_owner.php</file>
             <file>shared_tests/create_gene_IVD.php</file>
             <file>shared_tests/create_disease_IVA.php</file>
             <file>authorization_tests/create_individual_diagnosed_with_IVA.php</file>
